adds paranoia check before final call to rm

+++ build.sh
@@ -51,4 +51,4 @@ make install
 popd
 
 # Cleanup
-rm -rf "$BUILDDIR"
+[[ -d $BUILDDIR ]] && rm -rf "$BUILDDIR"
